[TASK] Shorten language text for webhooks description
This is a change to ease translation of the existing
texts.

The phrase "for your internal reference" adds little value
to the description descripion and causes difficulties
while translating: The phrase is ambiguous, it can mean
"for your internal documentation" or something else
and I am not sure what it means in this context.

Resolves: #100299
Releases: main
Change-Id: I9b6583c9a5315eab3e7c048eb61c4c06c0d236a9
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78255
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Georg Ringer <georg.ringer@gmail.com>