[BUGFIX] Ensure correct type resolving in `AbstractSchemaManager`
Doctrine DBAL 4.x removed the feature to save specific type
information as part of the column comment and for reading
the database schema with `AbstractSchemaManager` hierarchy
classes.

That also includes the option to enable and disable this
feature along with setting in type classes which needs to
write the custom information in case used database do not
support it natively - or for implemented custom type.

Doctrine postulate that it is the responsibility of the
application or framework to keep track of a precise and
cross database schema on own behalf, where the provided
schema information retrieved with the Doctrine DBAL
`AbstractSchemaManager` implementation are types which
would produce the same database structure for that vendor.

The possible TYPO3 counterpart for expected field types is
`$GLOBALS['TCA']` (TcaSchemaFactory), but does not really
ensure to hold all information, because TYPO3 allows extension
authors to define tables not managed by TYPO3 (TCA) with
`ext_tables.sql` files.

With #99062 the introduction of the native JSON database field
support it has been postulated that simply providing an array
for `Connection` methods like `insert` or `update` and using
`DataHandler` for TCA-type JSON handles this for all databases,
even if not nativly supported. That broke in TYPO3 v13 due to
the breaking change of Doctrine.

This mitigate this issue, this change reintroduces column
comment based type handling by ...

* introducing `GetColumnDeclarationSQLCommentTypeAwareTrait`
  used within extended `AbstractPlatform` classes to ensure
  adding column comment type hints for types on platforms
  needing it.
* introducing `ColumnTypeCommentMethodsTrait` used within
  extended `AbstractSchemaManager` classes to ensure that
  the correct type is set when reading database structure
  information.
* modifying extended `AbstractSchemaManager` classes to use
  the trait methods in suitable places, partly needing to
  clone code from Doctrine DBAL classes.
* removeing `ConnectionMigrator` SQLite specific override
  code switching to `TextType` for unsupported `JsonType`
  and resolving now the comment and todo with this change.

[1] https://github.com/doctrine/dbal/blob/54be50e2f2ea4cdbbdee35b9d0cf08cd9cb9bcb8/UPGRADE.md#bc-break-deployed-database-schema-no-longer-contains-the-information-about-abstract-data-types
[2] https://github.com/doctrine/dbal/blob/54be50e2f2ea4cdbbdee35b9d0cf08cd9cb9bcb8/UPGRADE.md#bc-break-removed-platform-commented-type-api

Resolves: #105208
Resolves: #105004
Related: #105005
Related: #99062
Releases: main
Change-Id: I497f4f84635ec922401c4169b577e2669167d577
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86436
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Stefan Bürk <stefan@buerk.tech>