[BUGFIX] Drop padding reset styles for btn-link usages
The applied padding for btn-link usages are already reset to 0,
forcing the padding to 0 has no impact and can be removed.

The edit link for for a column was missing the btn-link styling,
this is now corrected so it has the correct focus styling in addition
to the inheritance for colors.

Resolves: #105606
Releases: main, 13.4
Change-Id: I96e0546e5d9804adfdc7c1b48a48e19d3d35172d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86996
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Franzke <ben@bnf.dev>