[BUGFIX] Fix pagination of reactions and webhooks module
This fixes the pagination implementations, used
in the reactions and the webhooks module, which
are basically the same code base.

When filtering the "all count" is now properly
calculated preventing that the pagination is
always displayed even if there are only a
subset of records available.

The pagination refresh button now actually
refreshes the current paginated page.

A duplicated records object is removed.

Resolves: #106272
Releases: main, 13.4, 12.4
Change-Id: Ib95214dbe39b8aded03ec84938c39b5eeddd1586
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88432
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Oliver Bartsch <bo@cedev.de>