[BUGFIX] Avoid ORDER BY from count queries in Webhooks/Reactions
PostgreSQL is stricter than MySQL when it comes
to `GROUP BY` clauses and `ORDER BY` in count
queries. When using `count(*)` with `ORDER BY`
on a column that's not in the `GROUP BY` clause,
PostgreSQL throws an error.

The `ORDER BY` clause is unnecessary for count
queries and can cause compatibility issues with
PostgreSQL. This change adds an optional parameter
to skip `ORDER BY` when counting records.

Resolves: #107815
Releases: main, 13.4
Change-Id: Id1184c8e266d813abfdf0ea9f7272214839c554d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91215
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Oli Bartsch <bo@cedev.de>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: Garvin Hicking <garvin@hick.ing>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Garvin Hicking <garvin@hick.ing>
Tested-by: Oli Bartsch <bo@cedev.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>